edited files for printf/ testing main function done

+++ PRINTF/ft_printf.h
@@ -19,10 +19,11 @@
 
 int	ft_printf(const char *s, ...);
 int	ft_printf_char(char c);
-int	ft_printf_hexlow(unsigned int num);
-int	ft_printf_hexupp(unsigned int num);
+int	ft_printf_hexlow(unsigned long long num);
+int	ft_printf_hexupp(unsigned long long num);
 int	ft_printf_nbr(int n);
 int	ft_printf_string(char *s);
 int	ft_printf_unbr(unsigned int nb);
 int	ft_printf_pointer(void *ptr);
+
 #endif
